Start Notifications component refactor
Factor some things out of the mega Notifications component, and add a dummy
unit test to show willing
